Changed concurrency limit to depend on the active CPU count everywhere
