+++ Source/CarthageKit/Constants.swift
@@ -3,6 +3,10 @@ import Result
 
 /// A struct including all constants.
 public struct Constants {
+    
+    public static let concurrencyLimit: UInt = {
+        return UInt(ProcessInfo.processInfo.activeProcessorCount)
+    }()
 
     /// Carthage's bundle identifier.
     public static let bundleIdentifier: String = "org.carthage.CarthageKit"
